requirements.txt
humanfriendly==10.0
paho-mqtt==1.6.1
pyreadline3==3.4.1
